Create :preferences module and move DataStore there
