+++ core/preferences/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    id("atomtasks.android.library")
+    id("atomtasks.android.library.compose")
+    id("kotlin-android")
+    id("atomtasks.detekt")
+    id("atomtasks.android.library.jacoco")
+    id("atomtasks.android.hilt")
+}
+
+android {
+    namespace = "com.costular.atomtasks.core.preferences"
+}
+
+dependencies {
+    implementation(projects.core)
+    api(libs.preferences.datastore)
+}
