+++ settings.gradle.kts
@@ -36,3 +36,4 @@ include(":core:notifications")
 include(":core:logging")
 include(":baselineprofile")
 include(":core:review")
+include(":core:preferences")
